requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
pyserial==3.5
